xor-sum initial commit
Program which creates error correction xor (parity) similar to RAID5, with checksums added for each block to determine which block was corrupted.

It can only fix one xor block length of contiguous corruption. If more than 1 block has been affected, then it cannot recover the error. (So you need to specify a larger block size).

It's a very simple algorithm, and not as complicated as Error Correction Codes can go.